Refactoring pages
- Tidying up page structure before it gets out of hand
- Changed "ServerTalk" instances to use 'server'
- Adjusted ServerTalk.js usage to work with const pages
