+++ src/pages/CardCreator.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const CardCreator = () => {
+
+  return (
+
+    <div>
+        CardCreator
+    </div>
+
+  )}
+
+export default CardCreator
