+++ src/pages/Deck.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+const Deck = () => {
+
+  return (
+
+    <div>
+        Deck Page
+    </div>
+
+
+  )
+}
+
+export default Deck
