--- src/pages/deckview.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import Navbar from '../components/NavBar/NavBar.js'
-
-const deckview = () => {
-
-  return (
-
-    <div>
-        Deckview Page
-    </div>
-
-
-  )
-}
-
-export default deckview
