--- src/pages/newcard.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import Navbar from '../components/NavBar/NavBar.js'
-
-const newcard = () => {
-
-  return (
-
-    <div>
-        newcard page
-    </div>
-
-  )}
-
-export default newcard
